requirements.txt
streamlit == 1.51.0
pandas == 2.3.3
pytest == 8.4.2
requests == 2.32.5
fastapi == 0.120.1
pydantic == 2.12.3
uvicorn ==  0.38.0
mysql-connector-python == 9.5.0
